users: reformat user log view
Shaves 7 characters off for most lines, and gets rid of infamous
GAIN/LOSE display.

The terms "GAIN" and "LOSE" were originally introduced because having
negative numbers everywhere would look too, er, negative, and having a
"+" for positive numbers would get hard to notice in a right aligned
list. The visibility of the "+" was fixed a while ago, simply by adding
a space between the sign and the number, and now the same style is
applied to the user log view.

Old:
2023-07-05 06:24:54 LOSE   2   1.80 Example [2x 0.90]
2023-07-07 20:55:53 GAIN   1  20.00 Received from someone (example)

New:
2023-07-05 06:24:54     1.80 2x Example [2x 0.90]
2023-07-07 20:55:53  + 20.00 Received from someone (example)