v6.0.4: use readline's internal loop
This fixes the bug that empty lines would be inserted after each prompt,
starting from the first use of ^D.

Readline considers ^D end-of-file even when it's not, and for whatever
reason then adds a \n after BRACK_PASTE_FINI, which results in empty
lines after subsequent prompts.

With readline's internal loop, rl_found_eof gets reset to false, but
users of a custom loop don't get that luxury, and Term::ReadLine::Gnu
doesn't expose rl_found_eof (which was added to readline's API only a
few years ago) to do it manually.

One workaround, used briefly but not committed, would be to disable
bracketed paste.

A better workaround, as implemented by this commit, is to abandon the
custom loop and use readline's blocking call instead.